pyproject.toml
[tool.poetry]
name = "true-grit"
version = "0.1.0"
description = "A comparison of the 1969 and 2010 films."
authors = ["Rooster Cogburn <wrangler@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
behave = "^1.2.6"
dbt-duckdb = "^1.6.2"
ruff = "^0.1.3"
ipykernel = "^6.26.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
types-requests = "2.31.0.6"

[tool.poetry.group.dev.dependencies]
harlequin = "^1.1.1"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D211", "D213"]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["INP001", "D100", "D103", "ANN201"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
